backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
Pillow
docx2pdf
